Rename files and update book-detail
