+++ practice-one/src/detail.html
@@ -76,6 +76,6 @@
             </div>
         </form>
         <div class="overlay" id="overlay"></div>
-        <script type="module" src="./js/views/book-detail.js"></script>
+        <script type="module" src="./js/views/book-detail.view.js"></script>
     </body>
 </html>
