+++ practice-one/src/js/app.js
@@ -1,14 +1,13 @@
-import BookController from "./controller/controller";
-import BookView from "./views/view";
-import BookModel from "./models/model";
-
+import BookController from "./controller/book.controller";
+import BookView from "./views/book.view";
+import BookModel from "./models/book.model";
 
 export class App {
-    startApp() {
-        const bookModel = new BookModel();
-        const bookView = new BookView();
-        const bookController = new BookController(bookModel,bookView)
+  startApp() {
+    const bookModel = new BookModel();
+    const bookView = new BookView();
+    const bookController = new BookController(bookModel, bookView);
 
-        bookController.init();
-    };
-};
+    bookController.init();
+  }
+}
